오류해결 / restfull 로인하여 매핑의 문제가있었음. 추후 restfull 적응되면 변동될수있음.
