+++ src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -30,7 +30,7 @@
    </beans:bean>
 	
 	<resources mapping="/resources/**" location="/resources/" />
-	<resources mapping="**/resources/" location="/resources/" />
+	<resources mapping="**/resources/**" location="/resources/" />
 
 	<beans:bean class="org.springframework.web.servlet.view.InternalResourceViewResolver">
 		<beans:property name="prefix" value="/WEB-INF/views/" />
@@ -38,7 +38,6 @@
 	</beans:bean>
 	
 	<context:component-scan base-package="kr.co.*" />
-	
 	
 	
 </beans:beans>
