+++ src/main/webapp/WEB-INF/web.xml
@@ -52,7 +52,7 @@
 
 	<filter-mapping>
 		<filter-name>encodingFilter</filter-name>
-		<url-pattern>/*</url-pattern>
+		<url-pattern>/**</url-pattern>
 	</filter-mapping>
 
 </web-app>
